# Optimize `capitalizeWords()`