<!--
DO NOT DELETE THIS TEMPLATE.
Or at the very least make sure you include enough information so I know what you're adding.
-->

**Checklist:**

- [x] I've followed [the README contributing guidelines](https://github.com/Nerwyn/universal-remote-card/blob/main/.github/CONTRIBUTING.md) and ensured that pre-commit and build ran correctly.
- [x] I've done by best to ensure that my new code is stylistically similar to the rest of the project.
  - If it isn't too egregious the main developer will fix it in post.
- [x] I've tested that my changes work on my Home Assistant instance, on different browsers and webviews in needed.
- [x] I've ensured that my changes are not causing any new JavaScript errors in the browser console.
- [x] I've updated the editor to include new fields for new features if needed.
  - The editor is a mess and the main developer open to working on these changes if needed, but the ability to modify all configuration fields via the editor is a requirement to release the changes.
- [x] I've updated the README if needed.

**Type of Change:**

- [ ] Default keys, sources, and/or icons
- [ ] Bug fix
- [ ] New media platform
- [ ] Extending support for an existing media platform (i.e. keyboard support)
- [ ] New action (e.g. `perform-action`, `toggle`, `navigate`)
- [ ] New interaction (e.g. `tap_action`, `hold_action`)
- [x] Other

**Description of Changes**

<!--
Describe the code changes you've made and why you made them.
-->

Self-explanatory.

**Testing Methodology**

<!--
How did you test your changes?
Did you install the file to your Home Assistant server and ensure that the changes work?

If making default key/source/icon changes, did you test them like above or test them using the card's built in custom actions and icons? (This is fine for this kind of change)

What devices did you test your changes on? Web browsers?
-->

Compiled, replaced the bundle in my HA, and checked the tooltips. All working.

**Additional Information**

- Refs https://github.com/Nerwyn/universal-remote-card/commit/98d049d358f0ed9316e6ca4374765a7842b1d6b9#r157186810