@felipecrs 2025-05-15
I even found an issue with the previous function:

```
+ expect(capitalizeWords('hello_world_test')).toBe('Hello World Test');
    [Diff] Actual / Expected
-   Hello World_test
+   Hello World Test
```