@felipecrs 2025-05-15
Benchmark results:

```
benchmark               time/iter (avg)        iter/s      (min … max)           p75      p99     p995
----------------------- ----------------------------- --------------------- --------------------------
old capitalizeWords()            4.0 µs       249,400 (  3.5 µs … 196.4 µs)   3.7 µs   9.0 µs  17.8 µs
new capitalizeWords()            1.7 µs       580,900 (  1.7 µs …   2.0 µs)   1.7 µs   2.0 µs   2.0 µs

summary
  old capitalizeWords()
     2.33x slower than new capitalizeWords()
```

```ts
import { describe, it } from "jsr:@std/testing/bdd";
import { expect } from "jsr:@std/expect";
import { capitalizeWords } from './styles.ts';

Deno.bench({
    name: "old capitalizeWords()",
    baseline: true,
    fn() {
        function capitalizeWords(word: unknown) {
            return (word ?? '')
                .toString()
                .replace('_', ' ')
                .split(' ')
                .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
                .join(' ');
        }
        expect(capitalizeWords('hello')).toBe('Hello');
        expect(capitalizeWords('hello_world')).toBe('Hello World');
        expect(capitalizeWords('hello_world_test')).toBe('Hello World_test');
        expect(capitalizeWords('')).toBe('');
        expect(capitalizeWords(undefined)).toBe('');
        expect(capitalizeWords(123)).toBe('123');
        expect(capitalizeWords(null)).toBe('');
        expect(capitalizeWords(true)).toBe('True');
        expect(capitalizeWords('Hello_World')).toBe('Hello World');
        expect(capitalizeWords('hello__world')).toBe('Hello _world');
    }
})

Deno.bench({
    name: "new capitalizeWords()",
    fn() {
        expect(capitalizeWords('hello')).toBe('Hello');
        expect(capitalizeWords('hello_world')).toBe('Hello World');
        expect(capitalizeWords('hello_world_test')).toBe('Hello World Test');
        expect(capitalizeWords('')).toBe('');
        expect(capitalizeWords(undefined)).toBe('');
        expect(capitalizeWords(123)).toBe('123');
        expect(capitalizeWords(null)).toBe('');
        expect(capitalizeWords(true)).toBe('true');
        expect(capitalizeWords('Hello_World')).toBe('Hello World');
    }
})
```